Major layout refactoring
Implement LayoutMap context, which allows any child of a RegionView to
quickly query the location and size of any object in the region by ref.

Goal: ensure layout and rendering are ubiquitously cached, available to
any component that needs it, and is recalculated only when relevant data
is changed.

This was motivated by implementing z-fighting detection, which needs to
know not only every objects' position, but also its width, so it can
determine overlap.